apiser/10-grpc/requirement.txt
#aiocache==0.8.0
#aiofiles==0.3.2
#aioredis==1.1.0
async-timeout==2.0.1
#hiredis==0.2.0
httptools==0.0.10
#Jinja2==2.10
MarkupSafe==1.0
mongoengine==0.15.0
PyJWT==1.6.1
pymongo==3.6.1
#rediscache==0.0.1
sanic==0.7.0
Sanic-Cors==0.9.3
Sanic-Plugins-Framework==0.5.2.dev20180201
six==1.11.0
ujson==1.35
uvloop==0.9.1
websockets==4.0.1
